Address comments

+++ internal/helper/helper.go
@@ -413,6 +413,10 @@ func CheckTags(searchStruct interface{}, tag string) (string, error) {
 // BackupAndCopyResolvConf creates a backup of /etc/resolv.conf in a chroot
 // and copies the contents from the host system into the chroot
 func BackupAndCopyResolvConf(chroot string) error {
+	if osutil.FileExists(filepath.Join(chroot, "etc", "resolv.conf.tmp")) {
+		// already backed up/copied so do nothing
+		return nil
+	}
 	src := filepath.Join(chroot, "etc", "resolv.conf")
 	dest := filepath.Join(chroot, "etc", "resolv.conf.tmp")
 	if err := os.Rename(src, dest); err != nil {
@@ -422,6 +426,19 @@ func BackupAndCopyResolvConf(chroot string) error {
 	src = filepath.Join("/etc", "resolv.conf")
 	if err := osutil.CopyFile(src, dest, osutil.CopyFlagDefault); err != nil {
 		return fmt.Errorf("Error copying file \"%s\" to \"%s\": %s", src, dest, err.Error())
+	}
+	return nil
+}
+
+// RestoreResolvConf restores the resolv.conf in the chroot from the
+// version that was backed up by BackupAndCopyResolvConf
+func RestoreResolvConf(chroot string) error {
+	if osutil.FileExists(filepath.Join(chroot, "etc", "resolv.conf.tmp")) {
+		src := filepath.Join(chroot, "etc", "resolv.conf.tmp")
+		dest := filepath.Join(chroot, "etc", "resolv.conf")
+		if err := os.Rename(src, dest); err != nil {
+			return fmt.Errorf("Error moving file \"%s\" to \"%s\": %s", src, dest, err.Error())
+		}
 	}
 	return nil
 }
